Move Main components to Main folder
